python-services/face_recognition/requirements.txt
face_recognition==1.3.0
opencv-python==4.11.0.86
numpy==1.24.3
Pillow==10.0.0
python-socketio==5.8.0
aiohttp==3.8.5
python-engineio==4.5.1 